1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include "function_pointers.h"
+#include <stddef.h>
 /**
  * array_iterator - executes a function per array element
  * @array: array to iterate over
@@ -7,7 +7,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	size_t i;
 
 	if (action != NULL && array != NULL)
 		for (i = 0; i < size; i++)
